@@ -10,27 +10,31 @@ description: Babes-Bolyai
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: ">=2.2.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  css_colors:  
-  collection: 
-  url_launcher: 
+  share: ^0.6.2+1
+  link_text:
+  image_picker: ^0.6.1+4
+  path_provider: ^0.4.1
+  css_colors:
+  collection:
+  url_launcher:
   curved_navigation_bar: ^0.2.1
   connectivity:
   shimmer:
   transparent_image:
   http:
+  html:
   liquid_pull_to_refresh:
   rounded_floating_app_bar:
-  sorted_chips_row: #de implementat
-  flutter_swiper : 
-  flutter_sparkline: 
-  flutter_staggered_grid_view: 
-  flip_card:      #de vazut 
-  flutter_html_widget:
+  flutter_swiper:
+  flutter_sparkline:
+  flutter_staggered_grid_view:
+  flip_card: #de vazut
+  #flutter_html_widget:
   flutter_offline: "^0.2.4"
   flutter_debug_drawer:
   percent_indicator:
@@ -40,25 +44,34 @@ dependencies:
   mobile_popup:
   flutter_multi_carousel:
   shared_preferences:
+  flutter_form_builder: ^3.5.4
+  intl: ^0.16.0
+  eva_icons_flutter:
+  flutter_progress_button:
+  flutter_native_splash: #de implementat
+  progress_button:
+  rubber: # de observat
+  flutter_email_sender:
+  
+  #flutter_mailer:
   # local_auth:
-
   # tutorial_coach_mark:
- 
+  
+
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.2
+  flare_splash_screen: ^2.1.3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-
 
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
 
 # The following section is specific to Flutter.
 flutter:
-
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
@@ -66,30 +79,30 @@ flutter:
 
   # To add assets to your application, add an assets section, like this:
   assets:
-     - assets/svg/
-     - assets/images/
-     - assets/logo.png
-     - assets/alucard.jpg
-     - assets/tick.png
-     - assets/avatar.png
-     - assets/backdrop.png
-     - assets/video1_thumb.png
-     - assets/video4_thumb.png
-     - assets/video3_thumb.png
-     - assets/video2_thumb.png
+    - assets/TESTMYhtmlFILES/iNFO/cincisprezece.txt
+    - assets/Student UBB-black.flr
+    - assets/Student UBB-white.flr
+    - assets/svg/
+    - assets/images/
+    - assets/logo.png
+    - assets/alucard.jpg
+    - assets/tick.png
+    - assets/avatar.png
+    - assets/backdrop.png
+    - assets/video1_thumb.png
+    - assets/video4_thumb.png
+    - assets/video3_thumb.png
+    - assets/video2_thumb.png
   fonts:
-      - family: Nunito
-        fonts:
-          - asset: assets/Nunito.ttf
+    - family: Nunito
+      fonts:
+        - asset: assets/Nunito.ttf
   #  - images/a_dot_burr.jpeg
   #  - images/a_dot_ham.jpeg
-
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.io/assets-and-images/#resolution-aware.
-
   # For details regarding adding assets from package dependencies, see
   # https://flutter.io/assets-and-images/#from-packages
-
   # To add custom fonts to your application, add a fonts section here,
   # in this "flutter" section. Each entry in this list should have a
   # "family" key with the font family name, and a "fonts" key with a
