@@ -1,15 +1,16 @@
 import 'dart:ui';
 
+import 'package:UBB/Pages/Setari/widget/superhero.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import '../../../main.dart';
 import '../dummy/favourites.dart';
-import '../entities/favourite.dart';
 import '../utils/assets.dart';
 
 class FavouriteListView extends StatefulWidget {
   const FavouriteListView({Key key, @required this.onSelect}) : super(key: key);
 
-  final ValueChanged<Favourite> onSelect;
+  final ValueChanged<SuperHero> onSelect;
 
   @override
   State<StatefulWidget> createState() {
@@ -18,7 +19,6 @@ class FavouriteListView extends StatefulWidget {
 }
 
 class _FavouriteListViewState extends State<FavouriteListView> {
-
   ScrollController _favouriteScrollController = ScrollController();
 
   @override
@@ -38,30 +38,32 @@ class _FavouriteListViewState extends State<FavouriteListView> {
         controller: _favouriteScrollController,
         scrollDirection: Axis.horizontal,
         itemBuilder: (context, index) {
-          if (index == 0)
-            return Container(
-              margin: EdgeInsets.only(left: 20, bottom: 4, top: 20),
-              child: Image(image: Assets.image("add.png")),
-            );
-          else if (index == favourites.length + 1)
-            return Container(
-              width: 20,
-            );
-
-          Favourite favourite = Favourite().fromJson(favourites[index - 1]);
+          // if (index == 0)
+          //   return Container(
+          //     margin: EdgeInsets.only(left: 20, bottom: 4, top: 20),
+          //     child: Image(image: Assets.image("add.png")),
+          //   );
+          // else
+          // if (index == favourites.length+1)
+          //   return Container(
+          //     width: 20,
+          //   );
+          //!scos add-ul : "modificat itemcount + 2" si favorites[index-1]
+          // Favourite favourite = Favourite().fromJson(favourites[index]);
+          SuperHero yourProf = yourProfesors[index];
           return GestureDetector(
             onTap: () {
               _favouriteScrollController.animateTo(index * 84.0,
                   duration: Duration(milliseconds: 200), curve: Curves.linear);
-              widget.onSelect(favourite);
+              widget.onSelect(yourProf);
             },
             child: Card(
               clipBehavior: Clip.antiAliasWithSaveLayer,
-              margin: EdgeInsets.only(left: 20, bottom: 4, top: 20),
+              margin: EdgeInsets.only(left: 10,right:5, top: 20),
               shape: RoundedRectangleBorder(
                   borderRadius: BorderRadius.all(Radius.circular(16))),
               child: Image(
-                image: Assets.image(favourite.image),
+                image: NetworkImage(yourProf.photo),
                 fit: BoxFit.cover,
                 width: 64,
                 height: 64,
@@ -69,7 +71,7 @@ class _FavouriteListViewState extends State<FavouriteListView> {
             ),
           );
         },
-        itemCount: favourites.length + 2,
+        itemCount: yourProfesors.length,
       ),
     );
   }
